requirements.txt
blinker==1.9.0
click==8.1.8
fastapi
uvicorn[standard]
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
pymongo
motor
mysql-connector-python
python-multipart
starlette
jinja2
python-dotenv
pydantic-settings
